chore(supeset.utils.core): move all database utils to database utils module (#18058)
