+++ superset/examples/paris.py
@@ -19,15 +19,15 @@ import json
 import pandas as pd
 from sqlalchemy import inspect, String, Text
 
+import superset.utils.database as database_utils
 from superset import db
-from superset.utils import core as utils
 
 from .helpers import get_example_data, get_table_connector_registry
 
 
 def load_paris_iris_geojson(only_metadata: bool = False, force: bool = False) -> None:
     tbl_name = "paris_iris_mapping"
-    database = utils.get_example_database()
+    database = database_utils.get_example_database()
     engine = database.get_sqla_engine()
     schema = inspect(engine).default_schema_name
     table_exists = database.has_table_by_name(tbl_name)
