+++ superset/examples/sf_population_polygons.py
@@ -19,8 +19,8 @@ import json
 import pandas as pd
 from sqlalchemy import BigInteger, Float, inspect, Text
 
+import superset.utils.database as database_utils
 from superset import db
-from superset.utils import core as utils
 
 from .helpers import get_example_data, get_table_connector_registry
 
@@ -29,7 +29,7 @@ def load_sf_population_polygons(
     only_metadata: bool = False, force: bool = False
 ) -> None:
     tbl_name = "sf_population_polygons"
-    database = utils.get_example_database()
+    database = database_utils.get_example_database()
     engine = database.get_sqla_engine()
     schema = inspect(engine).default_schema_name
     table_exists = database.has_table_by_name(tbl_name)
