+++ tests/integration_tests/alerts_tests.py
@@ -22,6 +22,7 @@ from unittest.mock import patch
 import pytest
 from sqlalchemy.orm import Session
 
+import superset.utils.database
 from superset import db
 from superset.exceptions import SupersetException
 from superset.models.alerts import Alert, AlertLog, SQLObservation
@@ -56,7 +57,7 @@ logger = logging.getLogger(__name__)
 @pytest.yield_fixture(scope="module")
 def setup_database():
     with app.app_context():
-        example_database = utils.get_example_database()
+        example_database = superset.utils.database.get_example_database()
         example_database.get_sqla_engine().execute(
             "CREATE TABLE test_table AS SELECT 1 as first, 2 as second"
         )
@@ -88,7 +89,7 @@ def create_alert(
         recipients="recipient1@superset.com",
         slack_channel="#test_channel",
         sql=sql,
-        database_id=utils.get_example_database().id,
+        database_id=superset.utils.database.get_example_database().id,
         validator_type=validator_type,
         validator_config=validator_config,
     )
