+++ tests/integration_tests/databases/commands_tests.py
@@ -45,7 +45,8 @@ from superset.exceptions import (
     SupersetTimeoutException,
 )
 from superset.models.core import Database
-from superset.utils.core import backend, get_example_database
+from superset.utils.core import backend
+from superset.utils.database import get_example_database
 from tests.integration_tests.base_tests import SupersetTestCase
 from tests.integration_tests.fixtures.birth_names_dashboard import (
     load_birth_names_dashboard_with_slices,
