+++ tests/integration_tests/dict_import_export_tests.py
@@ -31,7 +31,7 @@ from superset.connectors.druid.models import (
     DruidCluster,
 )
 from superset.connectors.sqla.models import SqlaTable, SqlMetric, TableColumn
-from superset.utils.core import get_example_database
+from superset.utils.database import get_example_database
 from superset.utils.dict_import_export import export_to_dict
 
 from .base_tests import SupersetTestCase
