+++ tests/integration_tests/fixtures/birth_names_dashboard.py
@@ -26,7 +26,8 @@ from superset.connectors.sqla.models import SqlaTable
 from superset.models.core import Database
 from superset.models.dashboard import Dashboard
 from superset.models.slice import Slice
-from superset.utils.core import get_example_database, get_example_default_schema
+from superset.utils.core import get_example_default_schema
+from superset.utils.database import get_example_database
 from tests.common.example_data_generator.birth_names.birth_names_generator_factory import (
     BirthNamesGeneratorFactory,
 )
