+++ tests/integration_tests/fixtures/datasource.py
@@ -17,7 +17,8 @@
 """Fixtures for test_datasource.py"""
 from typing import Any, Dict
 
-from superset.utils.core import get_example_database, get_example_default_schema
+from superset.utils.core import get_example_default_schema
+from superset.utils.database import get_example_database
 
 
 def get_datasource_post() -> Dict[str, Any]:
