+++ tests/integration_tests/importexport/commands_tests.py
@@ -22,7 +22,7 @@ from freezegun import freeze_time
 
 from superset import security_manager
 from superset.databases.commands.export import ExportDatabasesCommand
-from superset.utils.core import get_example_database
+from superset.utils.database import get_example_database
 from tests.integration_tests.base_tests import SupersetTestCase
