+++ tests/integration_tests/queries/saved_queries/commands_tests.py
@@ -30,7 +30,7 @@ from superset.queries.saved_queries.commands.export import ExportSavedQueriesCom
 from superset.queries.saved_queries.commands.importers.v1 import (
     ImportSavedQueriesCommand,
 )
-from superset.utils.core import get_example_database
+from superset.utils.database import get_example_database
 from tests.integration_tests.base_tests import SupersetTestCase
 from tests.integration_tests.fixtures.importexport import (
     database_config,
