+++ tests/integration_tests/security_tests.py
@@ -40,9 +40,9 @@ from superset.models.slice import Slice
 from superset.sql_parse import Table
 from superset.utils.core import (
     backend,
-    get_example_database,
     get_example_default_schema,
 )
+from superset.utils.database import get_example_database
 from superset.views.access_requests import AccessRequestsModelView
 
 from .base_tests import SupersetTestCase
